blueprints/gke/hpc/bundle/app/images/video-download-api/requirements.txt
Flask==2.2.3
PyYAML==6.0
kubernetes==26.1.0
